Use schedutil governor by default
This is such a common default that there's no reason for Nerves not to
use it as well. It also removes one potential source of differences when
people see performance differences between Raspberry Pi OS and Nerves.

The article at https://lwn.net/Articles/682391/ describes the issues and
why schedutil is a good option unless there are overarching concerns
with a particular project.